+++ Mailer/app/Models/ContentGenerator/AnchorWirelinkReplace.php
@@ -7,6 +7,7 @@ use Nette\Http\Url;
 use Nette\InvalidArgumentException;
 use Remp\MailerModule\Models\ContentGenerator\GeneratorInput;
 use Remp\MailerModule\Models\ContentGenerator\Replace\IReplace;
+use Remp\MailerModule\Models\ContentGenerator\Replace\RtmClickReplace;
 
 class AnchorWirelinkReplace implements IReplace
 {
@@ -35,8 +36,19 @@ class AnchorWirelinkReplace implements IReplace
 
                 $quote = $matches[2][$idx] ?: '"';
 
+                $rtmClickHash = $url->getQueryParameter(RtmClickReplace::HASH_PARAM);
+                if ($rtmClickHash) {
+                    // we'll add rtm_click param to wrapper URL, no need to have it "inside"
+                    $hrefUrl = RtmClickReplace::removeRtmClickHash($hrefUrl);
+                }
+
                 $wirelinkUrl = new Url($this->wirelinkHost);
                 $wirelinkUrl->setPath(sprintf('/r/%s', rawurlencode($hrefUrl)));
+
+                if ($rtmClickHash !== null) {
+                    // set RTM click param to wrapped URL to make sure click tracking is able to extract it
+                    $wirelinkUrl->setQueryParameter(RtmClickReplace::HASH_PARAM, $rtmClickHash);
+                }
 
                 $href = sprintf('<a%shref=%s%s%s%s>', $matches[1][$idx], $quote, $wirelinkUrl, $quote, $matches[4][$idx]);
                 $content = str_replace($matches[0][$idx], $href, $content);